moved timeout test to conn_test
